+++ packages/browser-wallet/src/popup/pages/Account/i18n/en.ts
@@ -159,6 +159,9 @@ const t = {
             },
             noChanges: 'Your transaction contains no changes compared to the current baker.',
         },
+        remove: {
+            notice: 'Removing a baking will take effect after a cool-down period of {{cooldownPeriod}} days. During the cool-down period the baker stake is locked and cannot be changed.',
+        },
         details: {
             heading: 'Your baker is registered.',
             amount: 'Baker stake',
@@ -266,6 +269,9 @@ const t = {
                 title: 'Update during the cool-down period',
                 body: 'Only the delegation amount is locked during the cool-down period.\n\nThis means that you can still change restake status and target baker pool during the cool-down.',
             },
+        },
+        remove: {
+            notice: 'Removing delegation will take effect after a cool-down period of {{cooldownPeriod}} days. During the cool-down period the delegation amount is locked and cannot be changed.',
         },
         details: {
             heading: 'Your delegation is registered.',